refactor(core): Phase 1 — move DefaultStepExecutor + FlowOrchestratorBuilder to Core
Lifts two Hangfire-resident classes that had zero Hangfire dependencies into
FlowOrchestrator.Core, enabling future runtime adapters (queue-based, in-memory)
to share the same execution and configuration primitives without referencing Hangfire.

- DefaultStepExecutor: Hangfire/Execution → Core/Execution (promoted to public sealed)
- FlowOrchestratorBuilder: Hangfire/Configuration → Core/Configuration
  (HangfireEnabled made public getter so Hangfire assembly can read the flag)
- SqlServer, PostgreSQL, InMemory SC extensions updated to use Core namespace
- Tests updated to reflect new namespaces
