+++ src/FlowOrchestrator.InMemory/InMemoryServiceCollectionExtensions.cs
@@ -1,5 +1,5 @@
+using FlowOrchestrator.Core.Configuration;
 using FlowOrchestrator.Core.Storage;
-using FlowOrchestrator.Hangfire;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace FlowOrchestrator.InMemory;
